bug/IVYPORTAL-17514-Fix-GUI-tests-LE
- Update GUI tests
