+++ AxonIvyPortal/portal-selenium-test/src_test/com/axonivy/portal/selenium/common/BaseTest.java
@@ -17,6 +17,8 @@ import org.openqa.selenium.WindowType;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import com.axonivy.ivy.webtest.engine.EngineUrl;
+import com.axonivy.portal.selenium.page.NewDashboardPage;
+import com.axonivy.portal.selenium.page.UserProfilePage;
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.WebDriverRunner;
 
@@ -343,6 +345,13 @@ public class BaseTest {
       e.printStackTrace();
     }
     redirectToRelativeLink(String.format(createJSonFileUrl, filepath, key));
+  }
+  
+  public void resetFormattingLanguage() {
+    NewDashboardPage newDashboardPage = new NewDashboardPage();
+    UserProfilePage userProfilePage = newDashboardPage.openMyProfilePage();
+    userProfilePage.inputFormattingLanguage("English (United Kingdom)");
+    newDashboardPage = userProfilePage.save();
   }
 
   protected void showNewDashboard() {
