+++ AxonIvyPortal/portal-selenium-test/src_test/com/axonivy/portal/selenium/test/AnnouncementTest.java
@@ -1,5 +1,7 @@
 package com.axonivy.portal.selenium.test;
 
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -80,5 +82,10 @@ public class AnnouncementTest extends BaseTest {
     announcementPage.publish();
     announcementPage.dePublish();
     assertTrue(newDashboardPage.isAnnouncementMessageNotDisplayed());
+  }
+  
+  @AfterEach
+  public void resetDefaultFormattingLanguage() {
+    resetFormattingLanguage();
   }
 }
